# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trends Laravel API",
        "description": "企業の技術コミュニティでの影響力を分析・追跡するAPI",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/search/companies": {
            "get": {
                "tags": [
                    "検索"
                ],
                "summary": "企業検索",
                "description": "企業名・ドメイン・説明文から企業を検索します。",
                "operationId": "65205532c324c7142beb7954d74d90ba",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "検索クエリ（1-255文字）",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255,
                            "minLength": 1,
                            "example": "Google"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "取得件数（最大100）",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "検索結果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "companies": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "株式会社サンプル"
                                                            },
                                                            "domain": {
                                                                "type": "string",
                                                                "example": "sample.com"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "サンプル企業の説明"
                                                            },
                                                            "logo_url": {
                                                                "type": "string",
                                                                "example": "https://example.com/logo.png"
                                                            },
                                                            "website_url": {
                                                                "type": "string",
                                                                "example": "https://sample.com"
                                                            },
                                                            "is_active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "match_score": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 0.95
                                                            },
                                                            "current_rankings": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "total_results": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "search_time": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.123
                                                },
                                                "query": {
                                                    "type": "string",
                                                    "example": "サンプル"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "検索クエリが無効です"
                                        },
                                        "details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}